refactor: 페이징 처리에 필요한 RequestDto 통합
