+++ src/main/java/weavers/siltarae/mistake/dto/request/FeedRequest.java
@@ -3,28 +3,18 @@ package weavers.siltarae.mistake.dto.request;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
+import weavers.siltarae.global.dto.request.PageRequest;
 
 import static lombok.AccessLevel.PROTECTED;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class FeedRequest {
-    @NotNull(message = "페이지 번호는 필수 입니다.")
-    private Integer page;
-    @NotNull(message = "페이지 개수는 필수 입니다.")
-    private Integer size;
+public class FeedRequest extends PageRequest {
     @NotNull(message = "피드 타입은 필수 입니다.")
     private FeedType feedType;
 
     public FeedRequest(Integer page, Integer size, FeedType feedType) {
-        this.page = page;
-        this.size = size;
+        super(page, size);
         this.feedType = feedType;
-    }
-
-    public Pageable toPageable() {
-        return PageRequest.of(page, size);
     }
 }
